pyproject.toml
[project]
name = "aminder-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi>=0.124.2",
    "google-generativeai>=0.8.5",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
